+++ templates/web-angular/lib/app_component.dart
@@ -4,15 +4,26 @@
 import 'package:angular2/core.dart';
 import 'package:angular_components/angular_components.dart';
 
-import 'hello_dialog/hello_dialog.dart';
+import 'todo_list/todo_list_component.dart';
+import 'todo_list_service.dart';
 
 @Component(
   selector: 'my-app',
   styleUrls: const ['app_component.css'],
   templateUrl: 'app_component.html',
-  directives: const [materialDirectives, HelloDialog],
-  providers: const [materialProviders],
+  directives: const [materialDirectives, TodoListComponent],
+  providers: const [materialProviders, TodoListService],
 )
 class AppComponent {
-  // Nothing here yet. All logic is in HelloDialog.
+  String todoItemDescription = '';
+
+  @ViewChild('todoList')
+  TodoListComponent todoList;
+
+  bool get disableAdd => todoItemDescription.isEmpty;
+
+  void addItem() {
+    todoList.add(todoItemDescription);
+    todoItemDescription = '';
+  }
 }